Add debugging
