+++ src/composables/use-travel-records.ts
@@ -1,5 +1,7 @@
 import { TravelRecordService } from '@/services/travel-record-service';
+import { interpolateGreatCircle } from '@/utils/helper';
 import { useQuery } from '@tanstack/vue-query';
+import type { Position } from 'geojson';
 import { computed } from 'vue';
 
 export default function useTravelRecords() {
@@ -10,15 +12,37 @@ export default function useTravelRecords() {
     refetchOnReconnect: false,
   });
 
-  const travelRecords = computed(() => {
-    return data.value?.data?.map((travelRecord) => ({
+  const travelRecords = computed(() =>
+    data.value?.data?.map((travelRecord) => ({
       departure: travelRecord.departure?.location,
       destination: travelRecord.destination?.location,
-    }));
+    })),
+  );
+
+  const travelRecordGeoJson = computed(() => {
+    const coordinates: Position[][] = [];
+    travelRecords.value?.forEach((record) => {
+      coordinates.push(
+        ...interpolateGreatCircle(
+          [record.departure?.longitude ?? 0, record.departure?.latitude ?? 0],
+          [record.destination?.longitude ?? 0, record.destination?.latitude ?? 0],
+          20,
+        ),
+      );
+    });
+
+    return {
+      type: 'Feature',
+      geometry: {
+        type: 'MultiLineString',
+        coordinates: coordinates,
+      },
+      properties: {},
+    } as any;
   });
 
   return {
     isFetching,
-    travelRecords
+    travelRecordGeoJson,
   };
 }
